Make human member method
